requirements.txt
# Mosaic MCP Server & Agent
httpx>=0.27.0
pydantic>=2.0.0
mcp>=1.0.0
pydantic-ai>=0.1.0
python-dotenv>=1.0.0
logfire>=0.50.0
